feat(zinkup): move zinkup to the top level
