--- cli/Cargo.toml
@@ -1,41 +1,0 @@
-[package]
-name = "zinkup"
-description = "Zink toolchain"
-documentation = "https://docs.rs/zinkup"
-version.workspace = true
-authors.workspace = true
-edition.workspace = true
-license.workspace = true
-homepage.workspace = true
-repository.workspace = true
-
-[[bin]]
-path = "bin/elko.rs"
-name = "elko"
-required-features = ["elko"]
-
-[[bin]]
-path = "bin/zinkc.rs"
-name = "zinkc"
-required-features = ["zinkc"]
-
-[dependencies]
-anyhow.workspace = true
-cargo_metadata = { workspace = true, optional = true }
-clap = { workspace = true, features = ["derive"] }
-color-eyre.workspace = true
-colored = { workspace = true, optional = true }
-etc = { workspace = true, optional = true }
-semver = { workspace = true, optional = true }
-serde = { workspace = true, optional = true, features = [ "derive" ] }
-thiserror = { workspace = true, optional = true }
-toml = { workspace = true, optional = true }
-tracing.workspace = true
-tracing-subscriber = { workspace = true, features = ["std", "env-filter"] }
-wasm-opt = { workspace = true, optional = true }
-zinkc = { workspace = true, optional = true }
-
-[features]
-default = ["elko", "zinkc"]
-elko = [ "cargo_metadata", "colored", "etc", "semver", "serde", "thiserror", "toml", "wasm-opt", "zinkc" ]
-zinkc = [ "dep:zinkc" ]
