--- cli/README.md
@@ -1,25 +1,0 @@
-# `zinkup`
-
-The zink components are gathered here, you can install all of the
-components directly with:
-
-```bash
-cargo install zinkup
-```
-
-For installing only specified binaries:
-
-```bash
-cargo install zinkup --features elko,zinkc
-```
-
-Available binaries:
-
-| Name    | Description             |
-| ------- | ----------------------- |
-| `elko`  | Zink\'s package manager |
-| `zinkc` | The zink compiler       |
-
-## LICENSE
-
-GPL-3.0
