+++ utils/elko/src/lib.rs
@@ -1,0 +1,7 @@
+//! Zink package manager.
+
+mod build;
+mod new;
+pub mod utils;
+
+pub use self::{build::Build, new::New};
