+++ utils/elko/src/utils/manifest.rs
