+++ utils/elko/src/utils/result.rs
