+++ utils/elko/src/utils/wasm.rs
